# Owl v2 speedup