--- a/tests/inference/models_predictions_tests/test_owlv2.py
+++ b/tests/inference/models_predictions_tests/test_owlv2.py
@@ -1,5 +1,4 @@
 from inference.core.entities.requests.owlv2 import OwlV2InferenceRequest
-from inference.core.entities.responses.inference import ObjectDetectionInferenceResponse
 from inference.models.owlv2.owlv2 import OwlV2
 
 
@@ -25,57 +24,3 @@ def test_owlv2():
     # and of course the size of the model itself, ie base vs large
     # we set a tolerance of 1.5 pixels from the expected value, which should cover most of the cases
     assert abs(223 - response.predictions[0].x) < 1.5
-
-
-if __name__ == "__main__":
-    # run a simple latency test
-    image_via_url = {
-        "type": "url",
-        "value": "https://media.roboflow.com/inference/seawithdock.jpeg",
-    }
-
-    import requests
-    import tempfile
-    from PIL import Image
-    from inference.core.utils.image_utils import load_image_rgb
-    import time
-
-    # Download the image
-    response = requests.get(image_via_url["value"])
-    response.raise_for_status()
-
-    # Create a temporary file
-    with tempfile.NamedTemporaryFile(suffix=".jpg", delete=False) as temp_file:
-        temp_file.write(response.content)
-        temp_file_path = temp_file.name
-
-    img = Image.open(temp_file_path)
-    img = img.convert("RGB")
-    
-    request_dict = dict(
-        image=img,
-        training_data=[
-            {
-                "image": img,
-                "boxes": [{"x": 223, "y": 306, "w": 40, "h": 226, "cls": "post"}],
-            }
-        ],
-        visualize_predictions=False,
-    )
-
-    model = OwlV2()
-
-    for _ in range(10):
-        print("pre cache fill try")
-        time_start = time.time()
-        response = model.infer(**request_dict)
-        time_end = time.time()
-        print(f"Time taken: {time_end - time_start} seconds")
-
-        print("post cache fill try")
-        time_start = time.time()
-        response = model.infer(**request_dict)
-        time_end = time.time()
-        print(f"Time taken: {time_end - time_start} seconds")
-
-        model.reset_cache()